Fix discrepancy in compile_info
The BEAM disassembler used the atom 'none' to signify the absence
of a compile_info chunk in a .beam file. This clashes with the type
declaration of the compile_info field of a #beam_file{} record as
containing a list. Use [] to signify the absence of this chunk.
This simplifies the code and avoids a dialyzer warning.

For fixing a similar problem and for consistency, changed also the
return type of the attributes field of the #beam_file{} record.

This required a change in the beam_disasm test suite.